+++ RecordIN.cs
@@ -26,7 +26,7 @@ namespace Paint_Products_Database
         int stocksubmit;
         int tempamount;//
         int row;
-        int tempprice;//
+        double tempprice;//
         string strQuery;
 
 
@@ -261,7 +261,7 @@ namespace Paint_Products_Database
                     }
                     else
                     {
-                        string[] row = { txtID.Text, txtProductName.Text, cbxManufacturer.Text, cbxType.Text, txtStatus.Text, dateTimePicker1.Text, stocksubmit.ToString(), txtAmount.Text, (((Convert.ToInt32(txtStock.Text)) * (Convert.ToInt32(txtPrice.Text)))).ToString(), txtPrice.Text };
+                        string[] row = { txtID.Text, txtProductName.Text, cbxManufacturer.Text, cbxType.Text, txtStatus.Text, dateTimePicker1.Text, stocksubmit.ToString(), txtAmount.Text, (((Convert.ToInt32(txtStock.Text)) * (Convert.ToDouble(txtPrice.Text)))).ToString(), txtPrice.Text };
 
                         dataGridView2.Rows.Add(row);
 
@@ -382,7 +382,7 @@ namespace Paint_Products_Database
 
             tempamount = Convert.ToInt32(dataGridView2.CurrentRow.Cells["rAmount"].Value.ToString());
             stock = Convert.ToInt32(dataGridView2.CurrentRow.Cells["Stck"].Value.ToString());
-            tempprice = Convert.ToInt32(dataGridView2.CurrentRow.Cells["Prce"].Value.ToString());
+            tempprice = Convert.ToDouble(dataGridView2.CurrentRow.Cells["Prce"].Value.ToString());
 
 
             txtID.Text = dataGridView2.CurrentRow.Cells["ProductID"].Value.ToString();
